+++ internal/models/error.go
@@ -1,8 +1,8 @@
 package models
 
-type ErrorResponse struct {
-    Status  int    `json:"status"`
-    Message string `json:"message"`
-    Code    string `json:"code,omitempty"`
-    Details string `json:"details,omitempty"`
-}
+// type ErrorResponse struct {
+//     Status  int    `json:"status"`
+//     Message string `json:"message"`
+//     Code    string `json:"code,omitempty"`
+//     Details string `json:"details,omitempty"`
+// }
