+++ internal/repository/deparment_repository.go
@@ -1,4 +1,4 @@
-package repositories
+package repository
 
 import (
 	"database/sql"
